rewrote to use neo's built-in encoder and spark max pid controller

+++ src/main/java/frc/robot/constants/ArmConstants.java
@@ -4,9 +4,9 @@ import frc.robot.Robot;
 
 public class ArmConstants {
   public final int motorID = -1;
-  public final int kP = -1;
-  public final int kI = -1;
-  public final int kD = -1;
+  public final double kP = -1;
+  public final double kI = -1;
+  public final double kD = -1;
   public final double initialPosition = 0;
   public final double shelfPosition = -1;
   public final double intakePosition = -1;
